container-video-embeddings/04-retrieval/requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
boto3

awscli
botocore
langchain
faiss-cpu
psycopg2-binary
psycopg
langchain-aws
langchain-community
langchain_experimental
langchain_postgres

pypdf
numpy
Pillow
ffmpeg
tqdm

